Add a temporary custom error type for version key validation
The p2 codebase does not yet support the 'version' key within a
launchable stanza, which is a good thing to validate given that such a
launchable has no hope of being installed properly.

However, third party code that ultimately depends on this validation may
wish to ignore such an error in order to test p2-preparer builds that do
implement this behavior. Once version key support is merged, this new
error type can be removed.