replication: Fix health aggregation (#564)
The replication.AggregateHealth() call would return a podHealth object whose
health values would never change from those gathered in an initial query. As
used, the replication package would update a node and then only observe its
health from before the update, not after the update. This could allow a bad
update to be pushed to all nodes. The p2-replicate command and daemon set
updates are affected.

This bug happened because intended mutable shared state (a podHealth object)
was being passed by copy. The fix implemented in this commit is to pass by
reference.

There is also a potential for deadlock if the update thread encountered any
Consul error. It would send that error on a channel that was never read. This
mitigates the error by draining the channel.