Protect against empty strings in status keys in status store
The status store is designed to store status for many different resource
types within P2, each with their own structure. Therefore, the consul
key for a status is formatted as
"status/<resource_type>/<resource_id>/<namespace>" where resource type
is somthing like "pod_clusters", resource id is a uuid identifying which
pod cluster the status is for, and "namespace" is an extra identifier
allowing multiple interested sub systems to store their own status
structures independently of each other.

In anticipation of accidental changes using an empty resource type or
namespace having negative consequences on the structure of consul keys,
this commit enforces that each field is a non-empty string.