Add PodUniqueKey type.
This type is meant to identify a single pod instance. As of this commit,
this is always determined by combining the node hostname and pod id of a
running pod with a slash in that order, e.g. "example.com/mysql". This
is used in the intent tree and reality tree both.

Because this precludes P2 from being able to schedule multiple pods with
the same pod id on the same hostname, pods will begin to transition to
using UUIDs to uniquely identify them.

The type added in this commit is flexible enough to represent an "old"
key or a "new" uuid key. It also exposes a function for determining
whether the key is of the new or old format, which will be useful for
branching functionality in upcoming features.