Make ManifestResult return necessary information instead of key
Previously the consul key path was returned alongside the pod manifest
within a ManifestResult struct.

Rather than expose consul schema to every caller, it's better to
centralize the logic for parsing a node name and a PodUniqueKey from the
key.

The types package has a helper function for taking a consul key and
returning a PodUniqueKey type that knows if it follows the uuid or old
pod key spec.