Add consul implementation of podstore.Store interface
It currently supports Schedule(manifest, node),
Unschedule(PodUniqueKey), and Read(PodUniqueKey). Schedule and
unschedule attempt to maintain a secondary index for the pod at
/intent/<node>/<pod_uuid> referring to /pods/<uuid>, but some corner
cases will allow those to be out of sync.

A sweeper process is planned to delete indices or pods that do not have
their corresponding counterpart, which suggests a failed schedule or
deletion.