Make kp.Store handle uuid pods correctly
In a previous commit, the consulStore within the kp package was modified
to return PodUniqueKey and Node as a part of the ManifestResult struct.
In the case of a PodUniqueKey with .IsUUID == true, a UUIDNotImplemented
error was returned for these keys.

This commit implements pod uuid keys by looking them up as an index into
a pod store instance.