Mask *api.Client behind a ConsulClient interface in stores
This will allow providing a fake *api.Client to stores for use in tests
that provides all of the necessary functionality. The main motivation
for this commit is to add more tests to the kp.consulStore struct.

Unfortunately this change required a bit of gymnastics around Go's type
system. *api.Client is a struct with KV() and Session() methods that
return structs providing functionality for interacting with consul keys
and consul sessions, respectively.

The ConsulClient interface captures the KV() and Session() functions
that return sub-interfaces that we use, since tests will ultimately want
to fake out the KV and session portions of consul functionality.

Unfortunately, a raw *api.Client does not implement the ConsulClient
interface because the return types of its KV() and Session() functions
are structs, which is not an exact match to what the ConsulClient
interface requires. As a result, a consulutil.ConsulClientFromRaw()
function was added which returns an implementation of the ConsulClient
interface based on a passed *api.Client.

Since most consul client initialization code goes through
the kp.NewConsulClient function, this function simply calls the wrapper
function, so there shouldn't be much extra burden added to callers
initializing a store.