Add fake implementation of consulutil.ConsulKVClient
Additionally, unify the virtually identical Fake KV implementations in
consulutil and kptest. consulutil is preferred because kptest
necessarily depends on the kp package and the kp tests will now make use
of a fake KV.

Motivation:
The store types take a consulutil.ConsulKVClient as input to their
constructors, and the passed client is used to interact with consul. In
tests, it's convenient to mock out the consul functionality. This commit
adds a type with the KV functionality implemented that can stand in for
a ConsulKVClient, as well as makes a convenient function for
initializing one.