Fix race conditions when trying to update daemon sets
Previously, an update to daemon sets would quickly cancel the
replication and re-enact it, but it does not wait for replication to
gracefully release the session locks, this commit will fix that problem.

This will also remove the locks in daemon_set because they are not needed

They can be removed because each daemon set has their own instance of a
replication, and since all operations in a specific instance of a daemon
set are serialized, we can remove the locks.