Expose node name to Pod type
Add a "node" field to the "pods".Pod type to track the name of the node
the pod runs on. This is an expansion of the concept of a node, which
previously existed only for the preparer and other orchestration tools.
Lower-level tools that worked only worked on localhost now have a
"--node-name" flag to inform them of the node name to use if it differs
from the hostname. (The flag's name was chosen to avoid a conflict with
"--node" in many orchestration commands that specifies the node to
operate on.)