Allow the daemon set to act on changes when an error occurs
Previously, when the daemon set encounters an error during replication,
it would sleep for 5 minutes and then retry. It was not able to take in
updates or deletes until after the retry happens.

Now with this change it can take in updates in between this time period
and if the updates produce no error on replication it would cancel the
retry.