Include response output in remote API call errors (#589)
It is a hopeless task to debug a JSON decode error without knowing the
input document. Many JSON documents exist in P2 in a durable format
where they can be examined after a failures, e.g., a file on a disk or a
key in Consul. However, the response to a remote API call isn't always
reproducible, so this changes the error messages to include the first 80
characters of the response when querying the artifact registry and an
HTTP label applicator. I am guessing that 80 chars is enough to show
what kind of response was received without overwhelming the logs.