Add more env vars and convenience functions to hook env
This commit makes the hook env function for instantiating a pod struct
based on a pod installed on disk more explicit by renaming it to
PodFromDisk(). Furthermore, it exposes PodID(), PodHome(), and
PodConfigDirPath() functions in order to expose these variables without
needing to instantiate the full pod.

The PodConfigDirPath() function is backed by a newly exposed environment
variable pointing to the config directory (previously there was only a
variable pointing to the standard config file within that directory)

The idea is that this would prevent callers from accidentally depending
a pod being installed on the disk at a bad point in the lifecycle, for
example in a "before install" hook.