Add pods.PodFactory and pods.HookFactory interfaces
These interfaces can be used to create Pod structs for normal pods and
hooks, respectively.

The goal here is to reduce the number of places in the code base
computing the pod's home since this is going to become more complicated
when pod directories are namespaced by their UUID in consul. The
factories have one configurable variable specifying the filesystem root
to put pods under.