Make the roll farm robust to RUs with missing RCs.
If a rolling update exists in consul and is missing either its new or
old RC, the rolling update cannot be handled because the enable/disable
step at the beginning of the roll handler will fail.

This commit validates that both RCs in an RU exist before attempting to
run it. If one of the RCs is missing, the RU is deleted.