Remove buggy check in hookenv.PodFromDisk()
The PodFromDisk() function returned an error if there was no
current_manifest.yaml on the disk for that node. This was to protect
against hooks loading a pod struct from disk before it has been
installed.

However, this caused some problems with existing hooks so the change is
being rolled back.