Marshal cgroup memory as human-readable value
Prefer encoding "./pkg/util/size".ByteCount values, used in a
launchable's cgroup memory limit, into a human-readable format when
possible. It is expected that readable values are going to be common in
manifests, and it would be nice to preserve that readability when making
automatic manifest transformations.

This change will alter the SHA computation for a manifest. It also
breaks compatibility with older manifest schema versions that could not
parse readible values.