Remove IsUUID field from PodUniqueKey struct
The field was obviously used to differentiate PodUniqueKey structs that
signify a UUID from those that don't. Prior to this commit, the ID field
would be the consul key e.g. /intent/<node>/<pod_id> if IsUUID was
false.

After making use of PodUniqueKey, it's becoming clear that this struct
is ONLY useful in contexts where it's signifying a UUID. As a result,
this commit removes the IsUUID field and uses a nil pointer to a
PodUniqueKey to signify a legacy pod that does not have a UUID.