Use a unique pod name for file names instead of pod id
In the interest of allowing multiple pods with the same pod ID to run on
the same filesystem, we need to namespace files (e.g.
/etc/servicebuilder.d files) by the uuid of the pod running when
applicable.

To accomplish this, a new field uniqueName is added to the Pod struct.
If the pod was not created with a uuid, this uniqueName is just a copy
of the pod ID. If there is a uuid, the unique name is <pod_id>-<uuid>.