Fix RU bug if reality manifest is missing.
This commit fixes a bug that would occur when attempting to count the
number of healthy nodes in a replication controller if there is no
reality manifest for the pod in question.

This can easily occur if a pod has been scheduled for the first time
on a bad host which will result in the reality manifest remaining empty.