Add a rewritten WatchPrefix() function as WatchPrefixNew()
The WatchPrefix() function in the consulutil package has a control flow
that is not idiomatic in golang, and forces callers to employ bad
patterns in order to interact with it.

Specifically, WatchPrefix() has two output channels which it takes as
arguments in addition to a quit channel. As a result, the caller must
watch the quit channel as well and close the output channels.

WatchPrefixNew() creates a single output channel and returns it, and
pushes values onto the channel in a goroutine. The output channel is
closed when the goroutine exits due to the quit channel being closed.
This means callers of WatchPrefixNew() can simply exit when the output
channe is closed, for example using "range" inside of a "for".