Update pcstore.Watch() to use WatchPrefixNew()
This simplifies the structure of Watch() because the quit channel does
not need to be selected on in as many places. The goroutine will
naturally exit when the input channel from WatchPrefixNew() is closed.

However, this function will now block if the reader of the output
channel is slow or stops reading. There are no callers in this codebase.