daemon_set: skip locking
Now, InitializeDaemonSetReplication will pass `true` to the argument
controlling whether locking is skipped.
Resulting behavior change: Now, daemon sets will ignore locks.

This change is desired:
Consider the case where two daemon sets with the same pod ID are
started.
Currently, the second one will override the lock of the first one, and
the first one will stop replicating completely.
That means some nodes do not get the pod installed (if the second one
does not cover all desired nodes)

Since this is undesirable, instead we will skip locking, and trust that
daemon sets will have disjoint node selectors.
The consequence of this assumption being violated is that two daemon
sets would fight over control of any nodes that belong to both of them.

We know that current daemon sets are all disjoint, and currently daemon
set modifications are gated through us so we can maintain this as true.
In addition, we do perform a disjointness check at *creation* time (note
that labels can still change after daemon set creation such that some
node belongs to two daemon sets after they are both created).

The idea is that we will make this change and see if it becomes the case
that we have daemon sets that end up fighting, and take action then.