replication: Extract handle replication end from lock
lockHosts was spawning a goroutine that was nominally called
handleRenewalErrors but was doing more than handling renewal errors: It
was also handling the replication completion and cancellation.

handleRenewalErrors is renamed to handleReplicationEnd to better reflect
its functionality.

This renamed function is extracted from lockHosts, which is important
because an upcoming commit will cause the call to lockHosts to sometimes
be elided. We do want to call this new function even if we do not lock.