logging: Fix data race adding "Counter" to entries
ProcessCounter.Fire() contained a data race modifying the entry.Data map
to add the "Counter" and "PID" fields. Logrus makes a copy of the Entry
object, but it is shallow, so multiple loggers could share the same map.

The bug is fixed by having Fire() update a copy of Data.