consulutil: Fix deadlock when SessionManager exits
In SessionManager, once its "done" channel is closed, there is no
guarantee that any other goroutine will be reading its output channel.
The way to prevent blocking is to send inside a select statement. When a
session is acquired, SessionManager does this; when a session is lost or
released, that send wasn't guarded. As a consequence, the SessionManager
goroutine would never exit.

This problem is fixed by adding select around the second send.