consulutil: Add exponential backoff to SessionManager
Add an exponential backoff strategy to SessionManager to deal with
errors encountered while creating a session. Session loss is a
highly-correlated event, so jitter is added to mitigate a thundering
herd (mostly preparers) attempting to reestablish their sessions after a
loss occurs.

Errors encountered while renewing a session are probably a symptom of an
overloaded Consul server, but it isn't clear how to use an exponential
backoff to mitigate that problem. This changes uses a heuristic of
sending the next create request after the maximum timeout from the last
successful create.