rcstore: Add a CASDesiredReplicas function
Sometimes, AddDesiredReplicas isn't the desired amount of checking that
we would like to have. AddDesiredReplicas retries its mutation so that
it is resistant to *unrelated* changes (such as the changing of the
disable flag), but if someone else changes the desired replica count, we
want to know about it. Thus, a CAS for desired replicas.