rc: abolish rcmetrics in favor of p2metrics
We would like to add metrics on the number of RCs, RUs, and DSes. It
presents an inconsistent interface to callers if one of these (RCs)
requires a SetMetricsRegistry function and the others do not.

To head off the inconsistency, now we use the global p2metrics in the RC
Farm.