`cachedMatch` flag for `GetMatches`
This change introduces a new parameter to the `Applicator.GetMatches`
function, `cachedMatch`. It permits clients to specify whether or not
they tolerate cached results from the applicator or not.

The `ConsulApplicator` has been modified to use this flag to indicate
relying on underlying Consul aggreagator's cached label set, instead of
querying Consul for the full tree on each query.

The `HttpApplicator` has been altered to pass a `cachedMatch` query parameter
to the receiving service.

The ds farm implementation has been given a struct field that can
be set to enable stale queries for all daemon sets. All other usages
of `GetMatches` keep `cachedMatch` disabled for now.