Use byte instead of SHA equivalence to determine if ds manifest changed.
p2-dsctl will not update a daemon set if it perceives that no changes
have been made. In order to determine manifest equivalence, the SHA
of the existing and proposed manifests were being preparered. However,
manifest signature is not considered in SHA computation and as a result,
p2-dsctl did not allow updates that only change a manifest's signature,
which is sometimes an operational necessity.

Now, normalized byte equivalence of the manifests is used instead, so an
error will only be raised if the two manifests are exactly equivalent
including both contents and signature.