Share the mutex in fakeSession
It looks like FakePodStore is supposed to hold a single map ("locks")
and mutex to guard access to the map ("locksMu") that are both shared
among many fakeSession values. However, the existing code passed the
mutex by value. sync.Mutex values shouldn't ever be copied, and this
also undermines the intent of the fake stores.

This commit passes the mutex by reference so that fakeSessions can share
one value.