Fix the TestWatchAll test in dsstore package.
The test watches the daemon set channel returned by WatchAll, and the
test times out if a value is not sent on the channel after 5 seconds.
However, WatchAll() had a hardcoded 5 second time duration to wait after
receiving a daemon set update from WatchPrefix. This meant that the test
would always fail.

Now the pause time is passed in as a function argument, with a value of
0 used in the test so that it proceeds as fast as possible.