Add WriteRealityIndex() function to pod store.
This is used to write a PodUniqueKey index to the /reality tree. This
key can be used to find the pod in the pod store that is deployed.

The preparer will use this function to track what has and has not been
deployed, particularly so that uninstallation of a pod can occur by
deleting the /intent index. The preparer will see that a uuid exists in
/reality but not /intent and will uninstall as a result.