Reset pod installation backoff time if new manifest is received.
Commit f98b0a1116cc82caab27a6202d91aea4d0b1019f introduced an
exponential backoff for pod installation and launching when an error
occurs. The backoff starts at 1 second and doubles on each error. The
backoff time is only reset when a pod launch is successful.

This commit also resets the backoff time if a new manifest is received,
which will reduce the amount of time it requires for a "fixed" pod
manifest update to be processed. For example, if a manifest is deployed
with a bad launchable URL and the backoff time rises to a minute, it is
desirable that a corrected manifest can be scheduled and processed
immediately, instead of after a large backoff time.