Make types.PodUniqueKey just a string rather than a struct.
There is no utility to using it as a struct rather than a string.

For convenience, the pod factory now has two pod constructor functions,
NewLegacyPod() which behaves like the old NewPod() and initializes a pod
with an empty string PodUniqueKey. Another, NewUUIDPod() throws an error
if passed an empty string uuid.