Fix paths used by health watches
The current implementation of health watching does not correctly
account for services who have names that are prefixes of other services.

This means two services "crow" and "crowd" will both have health returned
for watches for the service "crow". This PR adds the node delimter in
explicitly to prevent this situation.