Disable each launchable before stopping any of them.
The previous behavior was to shut down with the following sequence:
1) disable 1st launchable
2) stop 1st launchable
3) disable 2nd launchable
...

The new behavior is to disable all launchables before stopping
launchables. This allows a launchable that depends on others for
successful operation to do shutdown handling prior to the other
launchables receiving term or kill signals.

Additionally, this commit fixes a bug where a failure in the bin/disable
script of a launchable would result in the launchable not being stopped
at all. Now it will be stopped regardless of exit status of bin/disable.