Modify PodStatus structure.
Three changes were made:

1) The word "service" was changed to "process" everywhere. Process more
precisely fits the notion of reporting process exit information.
2) Process "name" field is broken up into launchable id and entry point
name. P2 creates runit services with the naming schema of
"<pod_id>__<launchable_id>__<entry_point_script_name>", and we'd like to
avoid having callers understand this schema to locate the exit of the
particular process they care about.
3) ServiceState (aka process state) was removed because runit as the
process manager does not easily expose this information. It will be
practical to capture exit state of a process, and there is not yet a
clear use case for anything further than that.