Wire up PodProcessReporter to the preparer.
This commit adds preparer configuration for the PodProcessReporter,
namely path to the socket on which it should listen as well as the path
to an executable that will be invoked on process exit in order to
produce the JSON structured lines that the PodProcessReporter will
parse.

If the configuration is missing, no PodProcessReporter will be
instantiated or started.