add --extra-env flag to p2-exec
This flag makes it possible to pass an ENTRY_POINT env var to pod
processes. The env dir structure is insufficient because a launchable
may have multiple bin/launch files to be converted to runti services
which all share an env dir.

The ENTRY_POINT env var is necessary to report exit status for each
individual process in a launchable.