Emit warning messages if a launchable script is running for minutes.
Disable scripts that hang forever or run for a very long time are a
common source of confusion during p2 deploys, because they result in
p2-preparer appearing slow to accomplish the update. Emitting warning
messages to the logs makes it easier to identify when a slow launchable
lifecycle script is to blame.