Protect against empty label ID in labels.objectPath()
The labels package has a helper function objectPath(labelType, id) which
returns path.Join(labelType, id). In some cases this is fed to a
recursive consul delete request.

In theory any caller passing an empty ID could cause a deletion of the
entire label tree for that label type.

This commit ensures that a path cannot be computed using objectPath() if
id is the empty string.