Add integration test for process exit code making it to consul.
The test works as follows:

1) Schedule the p2-preparer with relevant config instructing it to write
finish files for each service it creates that will write exit
information to a SQLite database as well as poll that database
periodically to publish information to consul.
2) Schedule a hello pod as a uuid pod.
3) Use an HTTP request to instruct the hello process to exit with a
randomly chosen exit code.
4) Check that the exit code is reflected in the sqlite database within a
timeout.
5) Check that the exit code is reflected in consul within a timeout.

This commit also contains various improvements in robustness and
organization to the integration test.