Make restart policy a per-launchable setting instead of per-pod.
This will allow greater flexibility in the way pods are composed and
run. For example, the scheduled pod may have an entry point that should
only run once, such as a job like a database backfill. There may be
supporting launchables that need to be up to support the job, e.g. an
instance of stunnel which should be restarted if it crashes. However,
the job should not be restarted.