Attempt debug invocation of finish script in the event of failure.
If the hello service being killed does not result in a finish entry
being written to the sqlite database, it is difficult to debug the
cause.

This commit adds a direct invocation of the ./finish for the hello pod,
and logs the output and any associated errors. This will aid in
debugging the test if this is the case that failed.