Add configurable pause time between RC watches in RC farm.
The RC tree size can grow up to a few Mb in large deployments, which
means that running many instances of the RC farm results in a sizable
amount of bandwidth consumption.

This commit adds a configurable time.Duration to the RC farm which
causes it to pause in between watches on the entire RC tree. This trades
responsiveness to new replication controllers for a decrease in request
rate to the datastore (and thus a bandwidth reduction).

Individual RC watches are unaffected, so once an RC is noticed and
adopted by a farm, changes to replicas_desired will still be processed
very quickly.