artifact: Pass both pod ID and artifact name to discover endpoint
Instead of always passing only the launchable ID, we need to allow the
manifest to specify an override for the launchable.
An artifact_name field is added to the version stanza of a launchable
stana for this purpose.

At the same time, we need to also pass the pod ID to the discover
endpoint (and for this there is no override), so that we can check
whether the artifact<->pod pair makes sense.

So now the pod ID is passed as the parameter in the URL, while the
launchable ID is passed as a query parameter.