Fix bug in recovery from throttling of health publishing.
p2-preparer executes health checks on the pods it has deployed, and
writes the health value to consul if and only if that health value has
changed since the last time it was written. This is done to reduce QPS
of health checks on Consul.

This means that the preparer must keep a copy of its view of the health
value that is stored remotely in order to make the decision of whether
to write or not.

Additionally, services with flapping health will have their health
updates throttled. This is implemented by writing an "unknown" health
value to Consul, and ceasing all health writes until the service stops
flapping.

This commit fixes a bug where a throttle would cause "unknown" to be
written to Consul, however the local view of the remote health value
would be set to the actual service health (e.g. "passing" or
"critical"). As a result, when the service recovers from flapping, no
health will be written because the health value matches what the
preparer thinks is stored remotely.