Break out health update throttling to separate function.
This is an architectural but not a behavioral change to health updating.

Before this commit, there was one function called processHealthUpdater()
that receives health updates for a pod on a channel (e.g. 1 per second).
This function makes a lot of decisions:

1) only write an update to the datastore if the health has changed since
the last time we wrote health
2) if health is flapping between states at a rate greater than a
configured threshold, write "unknown" health to the datastore and wait a
configured timeout before publishing any more updates.

This commit adds a function throttleChecks() that handles most of the
logic for 2) above. It takes the raw health check stream as input and
returns a different health check stream that is now passed to
processHealthUpdater(). throttleChecks() takes responsibility for
handling throttling, and during a period of health flapping it will
write "unknown" values to the output channel instead of the underlying
health value, taking advantage of the fact that processHealthUpdater()
will not write the same health value twice.